% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_weather_option}
\alias{get_weather_option}
\title{Get weather package option}
\usage{
get_weather_option(name)
}
\arguments{
\item{name}{A character string specifying the name of the option to retrieve.}
}
\value{
The value of the specified option, or NULL if the option does not exist.
}
\description{
Retrieves the value of a specified option from the weather package configuration.
This function provides access to package-level settings and preferences.
}
\examples{
# Get the default frost threshold
get_weather_option("extreme.frost_threshold")

}
