% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{interpolation_function}
\alias{interpolation_function}
\title{Return a y value from a linear interpolation function}
\usage{
interpolation_function(x, y, values, split = "\\\\s+")
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{values}{values}

\item{split}{split}
}
\value{
The interpolated values
}
\description{
Return a y value from a linear interpolation function
}
