% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.r
\name{read_weather}
\alias{read_weather}
\title{Read weather records from a file list and/or a folder list}
\usage{
read_weather(file, format = "APSIM", ...)
}
\arguments{
\item{file}{A character string to specify weather filename.}

\item{format}{A character string to specify the format of weather file.}

\item{...}{Other arguments}
}
\value{
A data.frame which contains all weather data.
}
\description{
Read weather records from a file list and/or a folder list
}
\examples{
file <- system.file("extdata/ppd_72150.met", package = "tidyweather")
records <- read_weather(file)
head(records)
}
