% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{weather_reset}
\alias{weather_reset}
\title{Reset all tidyweather options to defaults}
\usage{
weather_reset()
}
\value{
Invisibly returns the default settings after reset.
}
\description{
This restores all package settings to their initial defaults as defined
when tidyweather was loaded. Use this if you want to undo all customizations.
}
\examples{
weather_options(extreme.frost_threshold = -2)
weather_reset()
weather_options()
}
