% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{day_length}
\alias{day_length}
\title{The time elapsed in hours between the specified sun angle
from 90 degree in am and pm. +ve above the horizon, -ve below the horizon.}
\usage{
day_length(doy, lat, angle = -6)
}
\arguments{
\item{doy}{day of year number}

\item{lat}{latitude of site (deg)}

\item{angle}{angle to measure time between, such as twilight (deg).
angular distance between 90 deg and end of twilight - altitude of sun. +ve up, -ve down.}
}
\value{
day length in hours
}
\description{
The time elapsed in hours between the specified sun angle
from 90 degree in am and pm. +ve above the horizon, -ve below the horizon.
}
