% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_distance.R
\name{spherical_distance}
\alias{spherical_distance}
\title{Calculate the sphere distance between two points on the Earth}
\usage{
spherical_distance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of the first point in degrees}

\item{lon1}{Longitude of the first point in degrees}

\item{lat2}{Latitude of the second point in degrees}

\item{lon2}{Longitude of the second point in degrees}
}
\value{
Distance in km
}
\description{
Calculate the sphere distance between two points on the Earth
}
\examples{
spherical_distance(34.05, -118.25, 40.7128, -74.0060) # Distance between Los Angeles and New York
}
