% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_time.R
\name{thermal_time}
\alias{thermal_time}
\title{Calculate thermal time using cardinal temperatures}
\usage{
thermal_time(mint, maxt, x_temp, y_temp, method = NULL, ...)
}
\arguments{
\item{mint}{The minimum temperature}

\item{maxt}{The maximum temperature}

\item{x_temp}{The cardinal temperatures}

\item{y_temp}{The effective thermal time}

\item{method}{The method to calculate thermal time.
The default method is ( maxt + mint ) / 2 - base.
The three hour temperature methods will be usesd if method = '3hr'.
The APSIM HourlySinPpAdjusted method can be selected with
method = 'HourlySinPpAdjusted'.}

\item{...}{Additional arguments passed to the method-specific functions.
Required arguments for method = 'HourlySinPpAdjusted' are, latitude and date.}
}
\value{
The thermal time.
}
\description{
Calculate thermal time using cardinal temperatures
}
\examples{
mint <- c(0, 10)
maxt <- c(30, 40)
x_temp <- c(0, 20, 35)
y_temp <- c(0, 20, 0)
thermal_time(mint, maxt, x_temp, y_temp)
thermal_time(mint, maxt, x_temp, y_temp, method = '3hr')
date <- as.Date(c("2020-01-01", "2020-01-02"))
thermal_time(mint, maxt, x_temp, y_temp, method = 'HourlySinPpAdjusted',
             latitude = -27.5, date = date)
}
